backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
requests==2.32.3

# Production server
gunicorn==23.0.0

# Static files serving
whitenoise==6.9.0
